WIP: first cut of weapons firing
It's rough around the edges at the moment.

- firing position and velocity is not quite aligned with the player
  polygon
- no bounce yet
- no expiry of projectiles
- no repeat rate control
- no damage or collision detection

It's also clear that we need some new concepts to avoid code duplication.

`apply_thrust` and `apply_edge_collisions` should work with Player and
Projectile.

Maybe a new module concept such as `PhysicalObject`?

There is also duplicate rendering code in PlayerLive that is crying out
to be behind a protocol or something.

I'm also toying with the idea of seperating the concerns of player
controls from application of player controls.